pinctrl: at91: add drive strength config options
The SAMA5 and SAMA92x5, SAM9G25 and SAM9G35 and others implement the ability to select the output drive strength for pins. The configuration typically allows low, medium and high power modes with predefined current limits. This patch reserves 3 bits in the pinctrl config(bits 5,6 and 7) for drive strength. The first bit enables setting the drive strength, the other two bits(6,7) set the drive strength to the specified level which is dependent on the chip.

i.e. A9s:
00 = high
01 = medium
10 = low
11 = reserved/undefined

the A5 does
00 = low
01 = low
10 = medium
11 = high

Signed-off-by: Marek Roszko <mark.roszko@gmail.com>