at91: clk: fix bug in divisor calculation
In at91sam9x5/sama5 family peripheral clock is produced by dividing MCK
by 2^div, where div=0..3. This patch fixes off-by-one error when div is set
to 4 when max frequency specified in devicetree is lower than MCK / 2^3.
This may happen only when system is misconfigured, but we still shouldn't
write illegal values to hardware, so div is set as to 3 and warning is printed
that the desired frequency is out of range.