Add/modify ethernet controler register definitions
This patch modifys and adds Atmel SAMA5D3x specific register definitions
so it matches the datasheet of the SAMA5D3x.
Previous definitions which are used in the corresponding macb.c file
are kept and marked with a comment (e.g. 'reserved') if they are not used
or used differently in the SAMA5D3x registers.