scripts: Makefile: generate symbols for DT nodes
In order to apply on a Device Tree an overlay that has a phandle to a
certain DT node, we need symbols in the DTB of the Device Tree to which
the overlay is applied.

Without the symbols, the tool in charge of applying the overlay cannot
find the DT node to apply the overlay to.

This adds the -@ option to the call to dtc which enables the generation
of symbols.

Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>