net: phy: micrel: Fix link-up failure at boot time
At boot time the link not come up if it was not already set by
U-Boot. The cable needed to be unplugged and then plugged again to
make the link come up.

This issue is caused by the Micrel KSZ9031 PHY which may fail to
establish a link when the Asymmetric Pause capability is set, as
described in a Silicon Errata (DS80000691D or DS80000692D), which
advises to always disable the capability.

This was fixed in mainline by 3aed3e2a143c ("net: phy: micrel: add
Asym Pause workaround"). But this patch doesn't apply on a 4.19 kernel
and need to be adapted to the old framework.

Signed-off-by: Gregory CLEMENT <gregory.clement@bootlin.com>