crypto: atmel-aes: Add fallback for unsupported ESP header length
Adding a software fallback for hmac(x)-cbc(aes) in cases where the ESP
header length is not a multiple of 4 words. The only ESP header length
that is not a multiple of 4 words is 6 words in this case, so this is the
only case tested.

As not all SoCs are affected by this limitation the HW version is used
to discriminate the SoCs. In this case HW versions 600 and 700 are
affected by this limitation.

To keep the hardware acceleration the check is put in place to use the
fallback for the highlighted case.

Signed-off-by: Ryan Wanner <Ryan.Wanner@microchip.com>
(cherry picked from commit 11e3595c5ed84cf86e960c79234df4a84f8e41f5)
Signed-off-by: Ryan Wanner <ryan.wanner@microchip.com>