drm/bridge: fix LVDS controller bus format
The current LVDS controller driver is hardcoded to map LVDS lanes to the
JEIDA format. Consequently, connecting an LVDS display that supports the
VESA format results in a distorted display due to the format mismatch.This
patch resolves the issue by querying the panel driver and setting the
appropriate format accordingly.

Signed-off-by: Sandeep Sheriker M <sandeep.sheriker@microchip.com>
Signed-off-by: Charan Pedumuru <charan.pedumuru@microchip.com>
Tested-by: Manikandan Muralidharan <manikandan.m@microchip.com>
Tested-by: Sandeep Sheriker M <sandeep.sheriker@microchip.com>
Reviewed-by: Dharma Balasubiramani <dharma.b@microchip.com>
[manikandan.m@microchip.com: Ported to 6.12.4, fixing conflicts]
Signed-off-by: Manikandan Muralidharan <manikandan.m@microchip.com>
[manikandan.m@microchip.com: Ported to 6.18.1, fixing conflicts]
Signed-off-by: Manikandan Muralidharan <manikandan.m@microchip.com>