ARM: dts: microchip: sama7g5: enable PIT64B for OP-TEE cases
Enable the PIT64B in the dts file to use them as the clocksource for
OP-TEE cases.

Currently TC0 is used as the clocksource and it can be affected when
changing the frequency of CPUPLL due to DVFS feature. This would
cause the failure of xtest 1006. Use PIT64B instead of TC0 can solve
the issue.

As this patch depends on the PIT64B driver in OP-TEE OS, please make
sure https://github.com/OP-TEE/optee_os/pull/7103 is applied when
use PIT64B as the clocksource.

Signed-off-by: Tony Han <tony.han@microchip.com>