gpio: mpfs: stop using parent hwirq number in interrupt handler
Currently gpio_index is calculated for the gpio driver using the
parent interrupt number, and doing a % 32 since the numbers are a
multiple of their index at the gpio controller when the interrupt parent
is the irq mux. That does not hold for coreGPIO when the parent
interrupt controller is actually the PLIC, and the method falls over
quickly. Revert to looping over all GPIO lines when one line triggers an
interrupt since this method "just works"TM without any need to calculate
which lines' interrupt was the trigger.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>