ARM: dts: microchip: sama7d65_curiosity: Enable SD UHS-I bus speed modes
The I/Os, between the SDMMC1 controller and the removable SD/SDIO device
inserted into the SDMMC1 slot, are part of the VDDSDMMC1 power domain
(vqmmc) which has a dual voltage capability (3.3V and 1.8V) on this board.
Selecting the proper signaling level, 3.3V or 1.8V, is achieved by toggling
the standard '1.8V Signaling Enable' bit (SDHCI_CTRL_VDD_180).

Since the underneath voltage switch is controlled as per the SDHC Standard,
there is no need to drive a standalone 'regulator-gpio'. Remove the
vqmmc-supply property.

Before the change, vqmmc-supply was indicating a fixed 3.3V signaling
level incompatible with UHS-I requirements.

The ability to power-cycle the SD device and toggle the signaling level
programmatically is required to negotiate UHS-I bus speed modes (SDR12 to
SDR104).

Signed-off-by: Aubin Constans <aubin.constans@microchip.com>
Acked-by: Nicolas Ferre <nicolas.ferre@microchip.com>