ARM: dts: microchip: sama5d2_xplained: Complete the e.MMC node
Augment the sdmmc0 node for the e.MMC wired to the SDMMC0 instance.

Assume the R225 resistor is populated, which is the factory default.
It pulls SDMMC0_1V8SEL down, which selects 3.3V for SAMA5D2's VDDSDMMC and
e.MMC's VCCQ power rails (the signaling voltage).

Continue to negotiate the High Speed DDR mode with the e.MMC, up to 52 MHz
depending on the Generic Clock.

Signed-off-by: Aubin Constans <aubin.constans@microchip.com>
Acked-by: Nicolas Ferre <nicolas.ferre@microchip.com>