pinctrl: at91-pio4: Add a pin polarity check for the wake up source
Add pin polarity check for the ULP1 wake up sources. There are cases
where the wake up source is active high. When the wake up source is
active high the WCR register in the PMC needs to be configured for
active high wake up source as well.

This fixes the missing polarity config for the wake up source.

Fixes: 908b7f056899 ("pinctrl: at91-pio4: Add GPIO wakeup for ULP1")
Signed-off-by: Ryan Wanner <Ryan.Wanner@microchip.com>
Tested-by: Manikandan Muralidharan <manikandan.m@microchip.com>